requirements.txt
asgiref==3.5.2
astroid==2.12.7
dill==0.3.5.1
Django==4.0
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.4.1
mccabe==0.7.0
Pillow==9.2.0
platformdirs==2.5.2
pylint==2.15.0
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
